feat : 배틀 UI 작업 진행
- 배틀 UI 작업 진행중

+++ Camp_FourthWeek(Basic_C#)/UIManager.cs
@@ -98,7 +98,7 @@ namespace Camp_FourthWeek_Basic_C__
             return (A.Item1+ B.Value.Item1, A.Item2);
         }
 
-        public static string GetBar(int cur, int Max)
+        public static string GetBar(float cur, float Max)
         {
             cur = Math.Max(0, Math.Min(cur, Max));
